requirements.txt
anyio==3.7.1
certifi==2025.8.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
openai==1.108.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.6
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.67.1
uvicorn==0.24.0
mysql-connector-python==8.3.0
pymysql==1.1.1
